+++ app/Http/Controllers/ContactController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Notifications\ContactNotification;
+use App\Rules\ReCaptchaValidation;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\Notification;
@@ -10,6 +11,15 @@ use Illuminate\Validation\Rule;
 
 class ContactController extends Controller
 {
+    /**
+     * @var \App\Rules\ReCaptchaValidation
+     */
+    protected $recaptchaValidation;
+
+    public function __construct(ReCaptchaValidation $recaptchaValidation) {
+        $this->recaptchaValidation = $recaptchaValidation;
+    }
+
     /**
      * Handle a contact page request for the application.
      *
@@ -51,7 +61,7 @@ class ContactController extends Controller
     protected function validator(array $data): \Illuminate\Validation\Validator
     {
         if (! isset($data['contactDetails'])) {
-            $data['contactDetails'] = ''; // could we skip this using some feature of the validator
+            $data['contactDetails'] = ''; // could we skip this using some feature of the validator?
         }
 
         $validSubjects = [
@@ -63,18 +73,12 @@ class ContactController extends Controller
         ];
 
         $validation = [
-            'subject'        => ['string', 'required', 'max:300', Rule::in($validSubjects)],
-            'name'           => ['string', 'required', 'max:300'],
-            'message'        => ['string', 'required', 'max:10000'],
-            'recaptcha'      => ['string', 'required', 'captcha'],
+            'recaptcha'      => ['required', 'string', 'bail', $this->recaptchaValidation],
+            'subject'        => ['required', 'string', 'max:300', Rule::in($validSubjects)],
+            'name'           => ['required', 'string', 'max:300'],
+            'message'        => ['required', 'string', 'max:10000'],
             'contactDetails' => ['string', 'nullable', 'max:300'],
         ];
-
-        // XXX: for phpunit dont validate captcha when requested....
-        // TODO this should be mocked in the test instead
-        if (getenv('PHPUNIT_RECAPTCHA_CHECK') == '0') {
-            unset($validation['recaptcha']);
-        }
 
         return Validator::make($data, $validation);
     }