+++ app/Providers/ReCaptchaServiceProvider.php
@@ -1,0 +1,40 @@
+<?php
+
+namespace App\Providers;
+
+use Illuminate\Support\ServiceProvider;
+use App\Rules\ReCaptchaValidation;
+use \ReCaptcha\ReCaptcha;
+
+class ReCaptchaServiceProvider extends ServiceProvider
+{
+   /**
+     * Register any application services.
+     *
+     * @return void
+     */
+    public function register()
+    {        
+        $this->app->bind(ReCaptchaValidation::class, function($app) {
+            $recaptcha = new ReCaptcha(
+                config('recaptcha.secret')
+            );
+
+            return new ReCaptchaValidation(
+                $recaptcha,
+                config('recaptcha.min_score'),
+                config('app.url')
+            );
+        });
+    }
+
+    /**
+     * Get the services provided by the provider.
+     *
+     * @return array
+     */
+    public function provides()
+    {
+        return [ReCaptchaValidation::class];
+    }
+}