+++ tests/Routes/User/RegisterTest.php
@@ -5,6 +5,7 @@ namespace Tests\Routes\User;
 use App\Invitation;
 use App\Notifications\UserCreationNotification;
 use App\User;
+use App\Rules\ReCaptchaValidation;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 use Illuminate\Support\Facades\Notification;
 use Tests\Routes\Traits\OptionsRequestAllowed;
@@ -19,20 +20,30 @@ class RegisterTest extends TestCase
 
     // TODO test password length when not deving
 
+    private function mockReCaptchaValidation($passes=true)
+    {
+        // replace injected ReCaptchaValidation class with mock (ContactController::$recaptchaValidation)
+        $mockRule = $this->createMock(ReCaptchaValidation::class);
+        $mockRule->method('passes')
+            ->willReturn($passes);
+    
+        $this->app->instance(ReCaptchaValidation::class, $mockRule);
+    }
+
     public function testCreate_Success()
     {
+        $this->mockReCaptchaValidation();
         Notification::fake();
 
         $invite = Invitation::factory()->create();
         $userToCreate = User::factory()->make();
 
-        putenv('PHPUNIT_RECAPTCHA_CHECK=0');
         $resp = $this->json('POST', $this->route, [
           'email' => $userToCreate->email,
           'password' => 'anyPassword',
           'invite' => $invite->code,
+          'recaptcha' => 'someToken'
         ]);
-        putenv('PHPUNIT_RECAPTCHA_CHECK=1');
 
         $resp->assertStatus(200)
         ->assertJsonStructure(['data' => ['email', 'id'], 'message', 'success'])
@@ -48,8 +59,11 @@ class RegisterTest extends TestCase
 
     public function testCreate_EmailAlreadyTaken()
     {
+        $this->mockReCaptchaValidation();
+
         $invite = Invitation::factory()->create();
         $user = User::factory()->create();
+
         $this->json('POST', $this->route, [
           'email' => $user->email,
           'password' => 'anyPassword',
@@ -61,8 +75,11 @@ class RegisterTest extends TestCase
 
     public function testCreate_NoInvitation()
     {
+        $this->mockReCaptchaValidation();
         $this->markTestSkipped('Fixme');
+
         $user = User::factory()->make();
+
         $this->json('POST', $this->route, [
           'email' => $user->email,
           'password' => 'anyPassword',
@@ -73,9 +90,11 @@ class RegisterTest extends TestCase
 
     public function testCreate_NoToken()
     {
-        putenv('PHPUNIT_RECAPTCHA_CHECK=1');
+        $this->mockReCaptchaValidation(false);
+
         $invite = Invitation::factory()->create();
         $user = User::factory()->make();
+
         $this->json('POST', $this->route, [
           'email' => $user->email,
           'password' => 'anyPassword',
@@ -87,6 +106,7 @@ class RegisterTest extends TestCase
 
     public function testCreate_NoEmailOrPassword()
     {
+        $this->mockReCaptchaValidation();
         $this->json('POST', $this->route, [])
         ->assertStatus(422)
         ->assertJsonStructure(['errors' => ['email', 'password']]);
@@ -94,7 +114,10 @@ class RegisterTest extends TestCase
 
     public function testCreate_BadInvitation()
     {
+        $this->mockReCaptchaValidation();
+
         $user = User::factory()->create();
+  
         $this->json('POST', $this->route, [
           'email' => $user->email,
           'password' => 'anyPassword',
@@ -106,7 +129,10 @@ class RegisterTest extends TestCase
 
     public function testCreate_BadEmail()
     {
+        $this->mockReCaptchaValidation();
+
         $invite = Invitation::factory()->create();
+        
         $this->json('POST', $this->route, [
           'email' => 'notAnEmail',
           'password' => 'anyPassword',